refactor: market_spot to spot v1
