+++ integrationtest/testdata/spot-v1/spot-quote-v1-kline.json
