+++ integrationtest/testdata/spot-v1/spot-quote-v1-ticker-24hr.json
