+++ integrationtest/testdata/spot-v1/spot-v1-symbols.json
